ZKR-2890-Refactor-Analyzer-From-Assignment
