+++ Cargo.toml
@@ -4,9 +4,5 @@ members = [
   "benchmarks"
 ]
 
-
-[patch."https://github.com/privacy-scaling-explorations/halo2.git"]
-halo2_proofs = { version = "0.2.0", git = "https://github.com/quantstamp/halo2.git" }
-
 [patch."https://github.com/privacy-scaling-explorations/halo2curves.git"]
 halo2curves = { git = "https://github.com/quantstamp/halo2curves.git",branch = "analyzer"}
