+++ korrekt/Cargo.toml
@@ -4,8 +4,8 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-# halo2_proofs = { git = "https://github.com/privacy-scaling-explorations/halo2.git", tag = "v2022_08_19" }
-halo2_proofs = { path = "../../halo2/halo2_proofs" }
+zcash_halo2_proofs = { git = "https://github.com/Analyzable-Halo2/zcash_halo2.git", package = "halo2_proofs", optional = true}
+pse_halo2_proofs = { git = "https://github.com/Analyzable-Halo2/pse-halo2.git", package = "halo2_proofs" , optional = true}
 gadgets = { git = "https://github.com/privacy-scaling-explorations/zkevm-circuits.git", rev = "2783eba338274a52036bf3db7fbc29575e24aa9f", package = "gadgets" }
 halo2curves = { git = 'https://github.com/privacy-scaling-explorations/halo2curves', tag = "0.2.1" }
 
@@ -16,8 +16,14 @@ anyhow = "1.0.71"
 regex = "1.8.4"
 num = "0.4.0"
 rayon = "1.5.1"
-ff = "0.12"
+ff = "0.13"
+group = "0.13"
 
 [lib]
 name = "korrekt"
 path = "src/lib.rs"
+
+[features]
+default = ["use_pse_halo2_proofs"]
+use_zcash_halo2_proofs = ["zcash_halo2_proofs"]
+use_pse_halo2_proofs = ["pse_halo2_proofs"]
