+++ korrekt/src/circuit_analyzer/abstract_expr.rs
@@ -1,5 +1,14 @@
-use halo2_proofs::{
-    arithmetic::FieldExt as Field,
+#[cfg(feature = "use_pse_halo2_proofs")]
+use pse_halo2_proofs::{
+    arithmetic::Field,
+    dev::CellValue,
+    plonk::{Advice, Any, Column, Expression, Selector,Phase},
+    poly::Rotation,
+};
+#[cfg(feature = "use_zcash_halo2_proofs")]
+use group::ff::Field;
+#[cfg(feature = "use_zcash_halo2_proofs")]
+use zcash_halo2_proofs::{
     dev::CellValue,
     plonk::{Advice, Any, Column, Expression, Selector},
     poly::Rotation,
@@ -21,13 +30,23 @@ pub enum AbsResult {
 /// This function traverses an expression tree and extracts the columns and rotations used within the expression.
 /// It recursively examines the expression and adds any encountered `Expression::Advice` columns and their corresponding rotations
 /// to the resulting set.
+#[cfg(any(feature = "use_zcash_halo2_proofs", feature = "use_pse_halo2_proofs",))]
 pub fn extract_columns<F: Field>(expr: &Expression<F>) -> HashSet<(Column<Any>, Rotation)> {
     fn recursion<F: Field>(dst: &mut HashSet<(Column<Any>, Rotation)>, expr: &Expression<F>) {
         match expr {
+            #[cfg(feature = "use_zcash_halo2_proofs")]
             Expression::Advice(advice_query) => {
                 let column = Column {
                     index: advice_query.column_index,
                     column_type: Advice {},
+                };
+                dst.insert((column.into(), advice_query.rotation));
+            }
+            #[cfg(feature = "use_pse_halo2_proofs")]
+            Expression::Advice(advice_query) => {
+                let column = Column {
+                    index: advice_query.column_index,
+                    column_type: Advice{ phase: advice_query.phase },
                 };
                 dst.insert((column.into(), advice_query.rotation));
             }
@@ -54,6 +73,7 @@ pub fn extract_columns<F: Field>(expr: &Expression<F>) -> HashSet<(Column<Any>, 
 /// It recursively traverses the expression tree and applies the corresponding evaluation rules to determine the result.
 /// The abstract result can be one of the following: `AbsResult::Zero`, `AbsResult::NonZero`, or `AbsResult::Variable`.
 ///
+#[cfg(any(feature = "use_zcash_halo2_proofs", feature = "use_pse_halo2_proofs",))]
 pub fn eval_abstract<F: Field>(
     expr: &Expression<F>,
     selectors: &HashSet<Selector>,
@@ -82,7 +102,8 @@ pub fn eval_abstract<F: Field>(
 
             let mut t = 0;
             if let CellValue::Assigned(fixed_val) = fixed[col][row] {
-                t = fixed_val.get_lower_128();
+                //t = fixed_val.get_lower_128();
+                t  = u64::from_str_radix(format!("{:?}",fixed_val).strip_prefix("0x").unwrap(), 16).unwrap();
             }
             if t == 0 {
                 AbsResult::Zero
@@ -122,5 +143,7 @@ pub fn eval_abstract<F: Field>(
                 eval_abstract(expr, selectors,region_begin,region_end,row_num,fixed)
             }
         }
+        #[cfg(feature = "use_pse_halo2_proofs")]
+        Expression::Challenge(_) => todo!(),
     }
 }
