+++ korrekt/src/sample_circuits/lookup_circuits/lookup_underconstrained.rs
@@ -1,222 +1,239 @@
-use halo2_proofs::{arithmetic::FieldExt, circuit::*, plonk::*, poly::Rotation};
-use std::marker::PhantomData;
+// #[cfg(feature = "use_pse_halo2_proofs")]
+// use pse_halo2_proofs::arithmetic::Field;
+// #[cfg(feature = "use_pse_halo2_proofs")]
+// use pse_halo2_proofs::circuit::*;
+// #[cfg(feature = "use_pse_halo2_proofs")]
+// use pse_halo2_proofs::plonk::*;
+// #[cfg(feature = "use_pse_halo2_proofs")]
+// use pse_halo2_proofs::poly::Rotation;
+// #[cfg(feature = "use_zcash_halo2_proofs")]
+// use group::ff::Field;
+// #[cfg(feature = "use_zcash_halo2_proofs")]
+// use zcash_halo2_proofs::circuit::*;
+// #[cfg(feature = "use_zcash_halo2_proofs")]
+// use zcash_halo2_proofs::plonk::*;
+// #[cfg(feature = "use_zcash_halo2_proofs")]
+// use zcash_halo2_proofs::poly::Rotation;
 
-#[derive(Debug, Clone)]
-pub struct FibonacciConfig {
-    pub advice: [Column<Advice>; 3],
-    pub s_add: Selector,
-    pub s_xor: Selector,
-    pub xor_table: [TableColumn; 3],
-    pub instance: Column<Instance>,
-}
+// use std::marker::PhantomData;
 
-#[derive(Debug, Clone)]
-struct FibonacciChip<F: FieldExt> {
-    config: FibonacciConfig,
-    _marker: PhantomData<F>,
-}
+// #[derive(Debug, Clone)]
+// pub struct FibonacciConfig {
+//     pub advice: [Column<Advice>; 3],
+//     pub s_add: Selector,
+//     pub s_xor: Selector,
+//     pub xor_table: [TableColumn; 3],
+//     pub instance: Column<Instance>,
+// }
 
-impl<F: FieldExt> FibonacciChip<F> {
-    pub fn construct(config: FibonacciConfig) -> Self {
-        Self {
-            config,
-            _marker: PhantomData,
-        }
-    }
+// #[derive(Debug, Clone)]
+// struct FibonacciChip<F: PrimeField> {
+//     config: FibonacciConfig,
+//     _marker: PhantomData<F>,
+// }
 
-    pub fn configure(meta: &mut ConstraintSystem<F>) -> FibonacciConfig {
-        let col_a = meta.advice_column();
-        let col_b = meta.advice_column();
-        let col_c = meta.advice_column();
-        let s_add = meta.selector();
-        let s_xor: Selector = meta.complex_selector();
-        let instance = meta.instance_column();
+// impl<F: PrimeField> FibonacciChip<F> {
+//     pub fn construct(config: FibonacciConfig) -> Self {
+//         Self {
+//             config,
+//             _marker: PhantomData,
+//         }
+//     }
 
-        let xor_table = [
-            meta.lookup_table_column(),
-            meta.lookup_table_column(),
-            meta.lookup_table_column(),
-        ];
+//     pub fn configure(meta: &mut ConstraintSystem<F>) -> FibonacciConfig {
+//         let col_a = meta.advice_column();
+//         let col_b = meta.advice_column();
+//         let col_c = meta.advice_column();
+//         let s_add = meta.selector();
+//         let s_xor: Selector = meta.complex_selector();
+//         let instance = meta.instance_column();
 
-        meta.enable_equality(col_a);
-        meta.enable_equality(col_b);
-        meta.enable_equality(col_c);
-        meta.enable_equality(instance);
+//         let xor_table = [
+//             meta.lookup_table_column(),
+//             meta.lookup_table_column(),
+//             meta.lookup_table_column(),
+//         ];
 
-        meta.create_gate("add", |meta| {
-            //
-            // col_a | col_b | col_c | selector
-            //   a      b        c       s
-            //
-            let s = meta.query_selector(s_add);
-            let a = meta.query_advice(col_a, Rotation::cur());
-            let b = meta.query_advice(col_b, Rotation::cur());
-            let c = meta.query_advice(col_c, Rotation::cur());
-            vec![s * (a + b - c)]
-        });
+//         meta.enable_equality(col_a);
+//         meta.enable_equality(col_b);
+//         meta.enable_equality(col_c);
+//         meta.enable_equality(instance);
 
-        meta.lookup("XOR_lookup", |meta| {
-            let s = meta.query_selector(s_xor);
-            let lhs = meta.query_advice(col_a, Rotation::cur());
-            let rhs = meta.query_advice(col_b, Rotation::cur());
-            let out = meta.query_advice(col_c, Rotation::cur());
-            vec![
-                (s.clone() * lhs, xor_table[0]),
-                (s.clone() * rhs, xor_table[1]),
-                (s * out, xor_table[2]),
-            ]
-        });
+//         meta.create_gate("add", |meta| {
+//             //
+//             // col_a | col_b | col_c | selector
+//             //   a      b        c       s
+//             //
+//             let s = meta.query_selector(s_add);
+//             let a = meta.query_advice(col_a, Rotation::cur());
+//             let b = meta.query_advice(col_b, Rotation::cur());
+//             let c = meta.query_advice(col_c, Rotation::cur());
+//             vec![s * (a + b - c)]
+//         });
 
-        FibonacciConfig {
-            advice: [col_a, col_b, col_c],
-            s_add,
-            s_xor,
-            xor_table,
-            instance,
-        }
-    }
+//         meta.lookup("XOR_lookup", |meta| {
+//             let s = meta.query_selector(s_xor);
+//             let lhs = meta.query_advice(col_a, Rotation::cur());
+//             let rhs = meta.query_advice(col_b, Rotation::cur());
+//             let out = meta.query_advice(col_c, Rotation::cur());
+//             vec![
+//                 (s.clone() * lhs, xor_table[0]),
+//                 (s.clone() * rhs, xor_table[1]),
+//                 (s * out, xor_table[2]),
+//             ]
+//         });
 
-    fn load_table(&self, mut layouter: impl Layouter<F>) -> Result<(), Error> {
-        layouter.assign_table(
-            || "xor_table",
-            |mut table| {
-                let mut idx = 0;
-                for lhs in 0..6 {
-                    for rhs in 0..6 {
-                        table.assign_cell(
-                            || "lhs",
-                            self.config.xor_table[0],
-                            idx,
-                            || Value::known(F::from(lhs)),
-                        )?;
-                        table.assign_cell(
-                            || "rhs",
-                            self.config.xor_table[1],
-                            idx,
-                            || Value::known(F::from(rhs)),
-                        )?;
-                        table.assign_cell(
-                            || "lhs ^ rhs",
-                            self.config.xor_table[2],
-                            idx,
-                            || Value::known(F::from(lhs ^ rhs)),
-                        )?;
-                        idx += 1;
-                    }
-                }
-                Ok(())
-            },
-        )
-    }
+//         FibonacciConfig {
+//             advice: [col_a, col_b, col_c],
+//             s_add,
+//             s_xor,
+//             xor_table,
+//             instance,
+//         }
+//     }
 
-    #[allow(clippy::type_complexity)]
-    pub fn assign(
-        &self,
-        mut layouter: impl Layouter<F>,
-        nrows: usize,
-    ) -> Result<AssignedCell<F, F>, Error> {
-        layouter.assign_region(
-            || "entire circuit",
-            |mut region| {
-                self.config.s_add.enable(&mut region, 0)?;
+//     fn load_table(&self, mut layouter: impl Layouter<F>) -> Result<(), Error> {
+//         layouter.assign_table(
+//             || "xor_table",
+//             |mut table| {
+//                 let mut idx = 0;
+//                 for lhs in 0..6 {
+//                     for rhs in 0..6 {
+//                         table.assign_cell(
+//                             || "lhs",
+//                             self.config.xor_table[0],
+//                             idx,
+//                             || Value::known(F::from(lhs)),
+//                         )?;
+//                         table.assign_cell(
+//                             || "rhs",
+//                             self.config.xor_table[1],
+//                             idx,
+//                             || Value::known(F::from(rhs)),
+//                         )?;
+//                         table.assign_cell(
+//                             || "lhs ^ rhs",
+//                             self.config.xor_table[2],
+//                             idx,
+//                             || Value::known(F::from(lhs ^ rhs)),
+//                         )?;
+//                         idx += 1;
+//                     }
+//                 }
+//                 Ok(())
+//             },
+//         )
+//     }
 
-                // assign first row
-                let a_cell = region.assign_advice_from_instance(
-                    || "1",
-                    self.config.instance,
-                    0,
-                    self.config.advice[0],
-                    0,
-                )?;
-                let mut b_cell = region.assign_advice_from_instance(
-                    || "1",
-                    self.config.instance,
-                    1,
-                    self.config.advice[1],
-                    0,
-                )?;
-                let mut c_cell = region.assign_advice(
-                    || "add",
-                    self.config.advice[2],
-                    0,
-                    || a_cell.value().copied() + b_cell.value(),
-                )?;
+//     #[allow(clippy::type_complexity)]
+//     pub fn assign(
+//         &self,
+//         mut layouter: impl Layouter<F>,
+//         nrows: usize,
+//     ) -> Result<AssignedCell<F, F>, Error> {
+//         layouter.assign_region(
+//             || "entire circuit",
+//             |mut region| {
+//                 self.config.s_add.enable(&mut region, 0)?;
 
-                // assign the rest of rows
-                for row in 2..nrows {
-                    b_cell.copy_advice(|| "a", &mut region, self.config.advice[0], row)?;
-                    c_cell.copy_advice(|| "b", &mut region, self.config.advice[1], row)?;
+//                 // assign first row
+//                 let a_cell = region.assign_advice_from_instance(
+//                     || "1",
+//                     self.config.instance,
+//                     0,
+//                     self.config.advice[0],
+//                     0,
+//                 )?;
+//                 let mut b_cell = region.assign_advice_from_instance(
+//                     || "1",
+//                     self.config.instance,
+//                     1,
+//                     self.config.advice[1],
+//                     0,
+//                 )?;
+//                 let mut c_cell = region.assign_advice(
+//                     || "add",
+//                     self.config.advice[2],
+//                     0,
+//                     || a_cell.value().copied() + b_cell.value(),
+//                 )?;
 
-                    let new_c_cell = if row % 2 == 0 {
-                        self.config.s_add.enable(&mut region, row)?;
-                        region.assign_advice(
-                            || "advice",
-                            self.config.advice[2],
-                            row,
-                            || b_cell.value().copied() + c_cell.value(),
-                        )?
-                    } else {
-                        self.config.s_xor.enable(&mut region, row)?;
-                        region.assign_advice(
-                            || "advice",
-                            self.config.advice[2],
-                            row,
-                            || {
-                                b_cell.value().and_then(|a| {
-                                    c_cell.value().map(|b| {
-                                        let a_val = a.get_lower_32() as u64;
-                                        let b_val = b.get_lower_32() as u64;
-                                        F::from(a_val ^ b_val)
-                                    })
-                                })
-                            },
-                        )?
-                    };
+//                 // assign the rest of rows
+//                 for row in 2..nrows {
+//                     b_cell.copy_advice(|| "a", &mut region, self.config.advice[0], row)?;
+//                     c_cell.copy_advice(|| "b", &mut region, self.config.advice[1], row)?;
 
-                    b_cell = c_cell;
-                    c_cell = new_c_cell;
-                }
+//                     let new_c_cell = if row % 2 == 0 {
+//                         self.config.s_add.enable(&mut region, row)?;
+//                         region.assign_advice(
+//                             || "advice",
+//                             self.config.advice[2],
+//                             row,
+//                             || b_cell.value().copied() + c_cell.value(),
+//                         )?
+//                     } else {
+//                         self.config.s_xor.enable(&mut region, row)?;
+//                         region.assign_advice(
+//                             || "advice",
+//                             self.config.advice[2],
+//                             row,
+//                             || {
+//                                 b_cell.value().and_then(|a| {
+//                                     c_cell.value().map(|b| {
+//                                         let a_val = u64::from_str_radix(format!("{:?}",a).strip_prefix("0x").unwrap(), 16).unwrap();//a.get_lower_32() as u64;
+//                                         let b_val = u64::from_str_radix(format!("{:?}",b).strip_prefix("0x").unwrap(), 16).unwrap();//b.get_lower_32() as u64;
+//                                         F::from(a_val ^ b_val)
+//                                     })
+//                                 })
+//                             },
+//                         )?
+//                     };
 
-                Ok(c_cell)
-            },
-        )
-    }
+//                     b_cell = c_cell;
+//                     c_cell = new_c_cell;
+//                 }
 
-    pub fn expose_public(
-        &self,
-        mut layouter: impl Layouter<F>,
-        cell: AssignedCell<F, F>,
-        row: usize,
-    ) -> Result<(), Error> {
-        layouter.constrain_instance(cell.cell(), self.config.instance, row)
-    }
-}
+//                 Ok(c_cell)
+//             },
+//         )
+//     }
 
-#[derive(Default)]
-pub struct MyCircuit<F>(pub PhantomData<F>);
+//     pub fn expose_public(
+//         &self,
+//         mut layouter: impl Layouter<F>,
+//         cell: AssignedCell<F, F>,
+//         row: usize,
+//     ) -> Result<(), Error> {
+//         layouter.constrain_instance(cell.cell(), self.config.instance, row)
+//     }
+// }
 
-impl<F: FieldExt> Circuit<F> for MyCircuit<F> {
-    type Config = FibonacciConfig;
-    type FloorPlanner = SimpleFloorPlanner;
+// #[derive(Default)]
+// pub struct MyCircuit<F>(pub PhantomData<F>);
 
-    fn without_witnesses(&self) -> Self {
-        Self::default()
-    }
+// impl<F: PrimeField> Circuit<F> for MyCircuit<F> {
+//     type Config = FibonacciConfig;
+//     type FloorPlanner = SimpleFloorPlanner;
 
-    fn configure(meta: &mut ConstraintSystem<F>) -> Self::Config {
-        FibonacciChip::configure(meta)
-    }
+//     fn without_witnesses(&self) -> Self {
+//         Self::default()
+//     }
 
-    fn synthesize(
-        &self,
-        config: Self::Config,
-        mut layouter: impl Layouter<F>,
-    ) -> Result<(), Error> {
-        let chip = FibonacciChip::construct(config);
-        chip.load_table(layouter.namespace(|| "lookup table"))?;
-        let out_cell = chip.assign(layouter.namespace(|| "entire table"), 4)?;
-        chip.expose_public(layouter.namespace(|| "out"), out_cell, 2)?;
+//     fn configure(meta: &mut ConstraintSystem<F>) -> Self::Config {
+//         FibonacciChip::configure(meta)
+//     }
 
-        Ok(())
-    }
-}
+//     fn synthesize(
+//         &self,
+//         config: Self::Config,
+//         mut layouter: impl Layouter<F>,
+//     ) -> Result<(), Error> {
+//         let chip = FibonacciChip::construct(config);
+//         chip.load_table(layouter.namespace(|| "lookup table"))?;
+//         let out_cell = chip.assign(layouter.namespace(|| "entire table"), 4)?;
+//         chip.expose_public(layouter.namespace(|| "out"), out_cell, 2)?;
+
+//         Ok(())
+//     }
+// }
+
