+++ korrekt/src/sample_circuits/lookup_circuits/mod.rs
@@ -1,3 +1,8 @@
 pub mod lookup;
+pub mod lookup_pse;
 pub mod lookup_underconstrained;
+pub mod lookup_underconstrained_zcash;
+pub mod lookup_underconstrained_pse;
 pub mod multiple_lookups;
+pub mod multiple_lookups_pse;
+pub mod multiple_lookups_zcash;
