+++ korrekt/src/test/mod.rs
@@ -1,2 +1,3 @@
 pub mod integration_tests;
+pub mod integration_tests_zcash;
 pub mod pse_test_row_diff_is_zero;
